・リファクタリング

+++ Assets/Script/GameManager.cs
@@ -9,7 +9,7 @@ public class GameManager : SingletonMonoBehaviour<GameManager> {
     public float m_time = 0;
 
     private GameObject m_player;
-    private int m_score = 0;
+    private float m_score = 0;
     private int m_wait = 0;
     private int m_speed = 0;
     private int m_jump = 0;
@@ -20,8 +20,6 @@ public class GameManager : SingletonMonoBehaviour<GameManager> {
     public GameObject Player {
         get { return m_player; }
     }
-
-    #endregion
 
     #region System Calls
